convert to using OnPendingBlockCommittedEvent
